requirements.txt
beautifulsoup4==4.13.4
docx2pdf==0.1.8
docx2txt==0.8
email_validator==2.2.0
fitz==0.0.1.dev2
fpdf==1.7.2
matplotlib==3.8.4
moviepy==1.0.3
pandas==2.2.3
Pillow>=9.0.0,<11
plotly>=5.23.0,<6
pushbullet.py==0.9.1
PyAutoGUI==0.9.54
PyPDF2==3.0.1
pytesseract==0.3.13
python_bcrypt==0.3.2
python_docx==1.1.2
reportlab==4.2.0
requests==2.32.3
seaborn==0.13.2
streamlit==1.38.0
streamlit-card==1.0.2
streamlit-extras==0.7.1
streamlit-js-eval==0.1.7
streamlit-lottie==0.0.5
streamlit-navigation-bar==3.3.0
streamlit-option-menu==0.3.13
streamlit-javascript==0.1.5
tools==1.0.2
rich<14
markdown-it-py==4.0.0
mdurl==0.1.2
pygments==2.19.2
twilio>=9.0.0
psycopg2-binary
sqlalchemy
supabase









